extranct withFetchingApp
