+++ pages/_app.tsx
@@ -2,14 +2,13 @@ import App, { Container } from 'next/app'
 import React from 'react'
 import { Provider } from 'mobx-react'
 import isServer from '../lib/isServer'
-import axios, { AxiosRequestConfig } from 'axios'
 import { CurrentUserState } from '../lib/CurrentUserState'
-import { createQueryMap } from '../lib/query'
 import { RouteState } from '../lib/RouteState'
 import JssProvider from 'react-jss/lib/JssProvider'
 import { MuiThemeProvider } from '@material-ui/core/styles'
 import { CssBaseline } from '@material-ui/core'
 import withPageContextApp from '../lib/withPageContextApp'
+import withFetchingApp from '../lib/withFetchingApp'
 import { PageContext } from '../lib/getPageContext'
 
 declare global {
@@ -26,66 +25,8 @@ export interface MyAppProps {
 }
 
 @withPageContextApp
+@withFetchingApp
 export default class MyApp extends App<MyAppProps> {
-  static async getPageProps (ctx, Component) {
-    return typeof Component.getInitialProps === 'function'
-      ? Component.getInitialProps(ctx)
-      : {}
-  }
-
-  static async getInitialProps ({ Component, router, ctx }) {
-    const {
-      pathname, query, asPath
-    } = ctx
-
-    let currentUser
-    let route
-    const options: AxiosRequestConfig = {}
-    if (isServer()) {
-      options.headers = ctx.req.headers
-      const { data } = (await axios.get('http://localhost:3000/auth', {
-        headers: ctx.req.headers
-      }))
-
-      if (data.currentUser) {
-        currentUser = new CurrentUserState(data.currentUser)
-      }
-
-      route = new RouteState({
-        pathname,
-        query: createQueryMap(query),
-        asPath
-      })
-    } else {
-      window.route.setRoute({
-        pathname,
-        query: createQueryMap(query),
-        asPath
-      })
-    }
-
-    return {
-      pageProps: await this.getPageProps(ctx, Component),
-      pathname, query, asPath,
-      currentUser,
-      route
-    }
-  }
-
-  componentWillMount () {
-    const {
-      currentUser,
-      route
-    } = this.props
-
-    if (!isServer()) {
-      if (currentUser) {
-        window.currentUser = new CurrentUserState(currentUser)
-      }
-      window.route = new RouteState(route)
-    }
-  }
-
   render () {
     const {
       Component,
